Refactor Qualifications component and remove unused code
