+++ src/app/(main)/apply/qualifications/Qualifications.tsx
@@ -2,13 +2,8 @@
 import { Application } from '@/app/(admin)/admin/applications/modals/Application';
 import { applicationsRepository } from '@/app/(admin)/admin/applications/repository';
 import { Certificate } from '@/app/(admin)/admin/certificates/Certificate';
-import { certificateRepository } from '@/app/(admin)/admin/certificates/repository';
 import {
-  Autocomplete,
-  AutocompleteItem,
   Button,
-  Divider,
-  Skeleton,
   Table,
   TableBody,
   TableCell,
@@ -17,9 +12,10 @@ import {
   TableRow,
 } from '@nextui-org/react';
 import { IconTrash } from '@tabler/icons-react';
-import React, { Key, useEffect } from 'react';
-import ResultsForm from './ResultsForm';
+import React from 'react';
 import { useApplication } from '../ApplicationProvider';
+import ResultsForm from './ResultsForm';
+import CertificateInput from './CertificateInput';
 
 export default function Qualifications() {
   const [certificate, setCertificate] = React.useState<Certificate>();
@@ -81,64 +77,5 @@ function ResultsTable({ application }: { application?: Application }) {
         ))}
       </TableBody>
     </Table>
-  );
-}
-
-type CertificateProps = {
-  setValue: (value: Certificate | undefined) => void;
-  value: Certificate | undefined;
-};
-
-function CertificateInput({ setValue, value }: CertificateProps) {
-  const [loading, setLoading] = React.useState(true);
-  const [certificates, setCertificates] = React.useState<Certificate[]>([]);
-  const application = useApplication();
-
-  useEffect(() => {
-    const other = {
-      courses: [],
-      gradingSchemes: [],
-      id: '-1',
-      name: 'Other',
-      passingGrade: null,
-    } as Certificate;
-    async function getData() {
-      try {
-        const data = await certificateRepository.getAll();
-        setCertificates([...data, other]);
-        const certificate = data.find(
-          (c) => c.id === application?.certificate?.id
-        );
-        setValue(certificate);
-      } finally {
-        setLoading(false);
-      }
-    }
-    getData();
-  }, [application?.certificate?.id, setValue]);
-
-  return (
-    <>
-      {loading ? (
-        <Skeleton className='w-full h-14 rounded-lg' />
-      ) : (
-        <Autocomplete
-          label='Highest Qualification'
-          className='w-full'
-          selectedKey={value?.id}
-          variant='bordered'
-          onSelectionChange={(item: Key) => {
-            setValue(certificates.find((c) => c.id === item));
-          }}
-          defaultItems={certificates}
-          description={!value && 'Please select your highest qualification'}
-          isLoading={loading}
-        >
-          {(item) => (
-            <AutocompleteItem key={item.id}>{item.name}</AutocompleteItem>
-          )}
-        </Autocomplete>
-      )}
-    </>
   );
 }
